Leverage non-lexical scoping to clean up example in Filesystem and Processes section
Non-lexical lifetimes were added in rust 1.31 (2018 edition) and 1.36 (2015 edition)
which have been available for over a year (1.36 released 4 July 2019). This small change
updates the code sample and removes a comment explaining the use of scoping to assist the
borrow checker. The removed comment was helpful to explain the reasoning behind the local
scoping, but was ultimately off topic for the section.

core::str<_>::trim_right was deprecated in Rust 1.33 and superceded by trim_end(). Update
file4.rs, and file5.rs to use trim_end.